+++ runtime/commands/default.js
@@ -431,26 +431,32 @@ Commands.hello = {
 
 Commands.setstatus = {
   name: 'setstatus',
-  help: 'This will change my current status to something else.',
-  usage: '<online / idle / twitch url> [playing status]',
+  help: 'Change my playing status on Discord to something else or pass nothing to clear the status!',
+  usage: '<online / idle / dnd / invisible / twitch url> [playing status]',
   level: 'master',
   fn: function (msg, suffix, bot) {
     var first = suffix.split(' ')
-    if (/^http/.test(first[0])) {
-      bot.User.setStatus(null, {
-        type: 1,
-        name: suffix.substring(first[0].length + 1),
-        url: first[0]
-      })
-      msg.channel.sendMessage(`Set status to streaming with message ${suffix.substring(first[0].length + 1)}`)
+    if (!suffix) {
+      bot.User.setStatus('online', null)
+      msg.channel.sendMessage(`Cleared status.`)
     } else {
-      if (['online', 'idle'].indexOf(first[0]) > -1) {
+      if (/^https?/.test(first[0])) {
+        bot.User.setStatus(null, {
+          type: 1,
+          name: (first[1] ? suffix.substring(first[0].length + 1) : null),
+          url: first[0]
+        })
+        msg.channel.sendMessage(`Set status to streaming with message ${suffix.substring(first[0].length + 1)}`)
+      } else if (['online', 'idle', 'dnd', 'invisible'].indexOf(first[0]) > -1) {
         bot.User.setStatus(first[0], {
-          name: suffix.substring(first[0].length + 1)
+          name: (first[1] ? suffix.substring(first[0].length + 1) : null)
         })
         msg.channel.sendMessage(`Set status to ${first[0]} with message ${suffix.substring(first[0].length + 1)}`)
+      } else if (suffix.substring(first[0].length + 1).length < 1) {
+        msg.reply('Can only be `online`, `idle`, `dnd` or `invisible`!')
       } else {
-        msg.reply('Can only be `online` or `idle`')
+        bot.User.setStatus('online', null)
+        msg.channel.sendMessage(`Cleared status.`)
       }
     }
   }
@@ -492,7 +498,7 @@ Commands['server-info'] = {
       }
       msg.channel.sendMessage('', false, embed)
     } else {
-      msg.channel.sendMessage("You can't do that in a DM, dummy!.")
+      msg.channel.sendMessage("You can't do that in a DM, dummy!")
     }
   }
 }