+++ runtime/commands/music.js
@@ -105,7 +105,7 @@ Commands.playlist = {
       } else {
         v.fetchList(msg).then((r) => {
           var arr = []
-          arr.push('Now playing: **' + r.info[0] + '** \n')
+          arr.push('Now playing: **' + r.info[0] + '** Requested by *' + r.requester[0] + '* \n')
           for (var i = 1; i < r.info.length; i++) {
             arr.push((i) + '. **' + r.info[i] + '** Requested by ' + r.requester[i])
             if (i === 9) {